requirements.txt
fastapi==0.110.0
uvicorn==0.28.1
jinja2==3.1.3
starlette==0.36.3

pydantic==2.6.4
pydantic-settings==2.2.1

psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
alembic==1.13.1

ruff==0.1.11
